+++ feature/main/src/main/java/com/example/main/MainScreenVM.kt
@@ -2,6 +2,7 @@ package com.example.main
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.domain.CourseState
 import com.example.domain.models.Course
 import com.example.domain.repositories.CourseRepository
 import com.example.ui.UiModule
@@ -10,16 +11,8 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
-
-data class CourseUiState(
-    val courses: List<Course> = emptyList(),
-    val isLoading: Boolean = false,
-    val errorMessage: String? = null
-)
-
 
 @HiltViewModel
 class MainScreenVM @Inject constructor(
@@ -27,58 +20,46 @@ class MainScreenVM @Inject constructor(
     private val stringProvider: UiModule.StringResourceProvider
 ): ViewModel() {
 
-    private val _uiState = MutableStateFlow(CourseUiState())
-    val uiState: StateFlow<CourseUiState> = _uiState.asStateFlow()
+    private val _state = MutableStateFlow<CourseState>(CourseState.Initial)
+    val state: StateFlow<CourseState> = _state.asStateFlow()
 
 
     fun loadData() {
-        viewModelScope.launch {
-            _uiState.update { it.copy(isLoading = true) }
+        if (_state.value is CourseState.Loading || _state.value is CourseState.Content)
+            return
 
+        _state.value = CourseState.Loading
+        viewModelScope.launch {
             try {
                 val data = courseRepository.getCourses(true).first()
-
-                _uiState.update {
-                    it.copy(
-                        isLoading = false,
-                        courses = data,
-                        errorMessage = null
-                    )
-                }
+                _state.value = CourseState.Content(data)
             } catch (e: Exception) {
-                _uiState.update {
-                    it.copy(
-                        isLoading = false,
-                        errorMessage = e.localizedMessage
-                            ?: stringProvider.getString(R.string.main_error_network)
-                    )
-                }
+                _state.value = CourseState.Error(e.localizedMessage ?: stringProvider.getString(R.string.main_error_network))
             }
         }
     }
 
     fun changeBookmarkOfCourse(course: Course) {
+        if (_state.value !is CourseState.Content)
+            return
+
         viewModelScope.launch {
             if (course.hasLike)
                 courseRepository.deleteCourse(course)
             else courseRepository.insertCourse(course.copy(hasLike = true))
         }.let {
-            val idx = _uiState.value.courses.indexOf(course)
-            _uiState.update {
-                it.copy(
-                    courses = it.courses.toMutableList().apply {
-                        this[idx] = course.copy(hasLike = !course.hasLike)
-                    }
-                )
-            }
+            val idx = (_state.value as CourseState.Content).courses.indexOf(course)
+            _state.value = CourseState.Content(
+                (_state.value as CourseState.Content).courses.toMutableList().apply {
+                    this[idx] = course.copy(hasLike = !course.hasLike)
+                }
+            )
         }
     }
 
     fun onSort() {
-        _uiState.update {
-            it.copy(
-                courses = it.courses.sortedByDescending { course -> course.publishDate },
-            )
-        }
+        _state.value = CourseState.Content((_state.value as CourseState.Content)
+            .courses.sortedByDescending { course -> course.publishDate }
+        )
     }
 }
